requirements.txt
scipy
tensorflow
seaborn
numpy
fbprophet
matplotlib
pandas
Pillow
scikit_learn
statsmodels
timesynth
